Initial commit.
At this point, what I have is a copy of `naga::back::wgsl` that has been
tinkered with until it produces syntactically valid Rust code, and
semantically valid code in some narrow cases. There is a lot of leftover
WGSLness, and after the original tinkering the code was roughly ripped
out into a separate crate and workspace so the organization is poor.
But it’s a place to start.