Add `no_std` to crates that need it.
Note that the `rt` crate does not currently build `no_std` due to
`glam`. We could fix the feature deps, but eventually we want
to replace `glam` with something custom which allows wider SIMD and
makes the code generator’s life easier, so I’m not going to fiddle
with `glam`’s features just yet.