Switch over to using `rt::Scalar` for scalars.
This causes arithmetic to be executed with Naga/shader semantics instead
of Rust language and standard library semantics (insofar as we implement
shader arithmetic correctly), and it will aid in adding explicit
vectorization support (because `Scalar` will be able to contain a vector
of subgroup/quad data).

The points at which translation between `Scalar<T>` and `T` occur are
not well-founded, and this change probably breaks some shader code that
previously worked; however, the tests pass and I believe it will be
easier to work forward from this than to seek immediate perfection.