Translate functions so that they have `impl Into` parameters.
This will clear the way for using `rt::Scalar` for scalars without
making the functions inconvenient to use by requiring explicit
conversion of everything.

I’m not yet sure whether `Into` is the right choice — it might be
better to produce strictly non-generic code — but it is flexible
while things are still changing.

Future work:
* Actually use `rt::Scalar` as planned.
* `use_into_for_arg()` should become part of some broader system of
  consistent type translation.
* Consider whether we really want `impl Into` or whether the shim
  functions should have merely different concrete types. (This question
  should be revisited once SIMD is working.)